+++ packages/api/test/pin.spec.js
@@ -4,18 +4,13 @@ import fetch from '@web-std/fetch'
 import { endpoint } from './scripts/constants.js'
 import { getTestJWT } from './scripts/helpers.js'
 import {
+  DATA_NOT_FOUND,
   ERROR_CODE,
-  ERROR_STATUS,
-  getEffectivePinStatus,
   INVALID_CID,
-  INVALID_META,
-  INVALID_NAME,
-  INVALID_ORIGINS,
-  REQUIRED_CID,
-  INVALID_LIMIT,
-  INVALID_REPLACE
+  INVALID_REPLACE,
+  getEffectivePinStatus
 } from '../src/utils/psa.js'
-import { PinningUnauthorizedError } from '../src/errors.js'
+import { PinningUnauthorizedError, PSAErrorResourceNotFound, PSAErrorInvalidData, PSAErrorRequiredData } from '../src/errors.js'
 
 /**
  *
@@ -106,7 +101,7 @@ describe('Pinning APIs endpoints', () => {
       assert.deepEqual(res.status, 401)
     })
 
-    it('validates filter values', async () => {
+    it('validates limit value passed as filter', async () => {
       const opts = new URLSearchParams({
         limit: '3.14'
       })
@@ -123,8 +118,26 @@ describe('Pinning APIs endpoints', () => {
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
       const error = await res.json()
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, INVALID_LIMIT)
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/limit: Instance type "number" is invalid. Expected "integer".')
+    })
+
+    it('requires status', async () => {
+      const url = new URL(`${baseUrl}`).toString()
+      const res = await fetch(
+        url, {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        })
+
+      assert(res, 'Server responded')
+      assert.strictEqual(res.status, ERROR_CODE)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorRequiredData.CODE)
+      assert.strictEqual(error.details, 'Instance does not have required property "status".')
     })
 
     it('validates CID values passed as filter', async () => {
@@ -146,13 +159,57 @@ describe('Pinning APIs endpoints', () => {
 
       assert.strictEqual(res.status, ERROR_CODE)
       const error = await res.json()
-      assert.strictEqual(error.reason, ERROR_STATUS)
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
       assert.strictEqual(error.details, INVALID_CID)
     })
 
-    it('returns only successful pins when no filter values are specified', async () => {
+    it('validates status values passed as filter', async () => {
+      const opts = new URLSearchParams({
+        status: 'pinning,badStatus'
+      })
+      const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
-        baseUrl, {
+        url, {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        })
+
+      assert.strictEqual(res.status, ERROR_CODE)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/status/1: Instance does not match any of ["queued","pinning","pinned","failed"].')
+    })
+
+    it('validates match values passed as filter', async () => {
+      const opts = new URLSearchParams({
+        match: 'badMatch'
+      })
+      const url = new URL(`${baseUrl}?${opts}`).toString()
+      const res = await fetch(
+        url, {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        })
+
+      assert.strictEqual(res.status, ERROR_CODE)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/match: Instance does not match any of ["exact","iexact","ipartial","partial"].')
+    })
+
+    it('returns only successful pins when no filter values are specified', async () => {
+      const opts = new URLSearchParams({
+        status: 'pinned'
+      })
+      const url = new URL(`${baseUrl}?${opts}`).toString()
+      const res = await fetch(
+        url, {
           method: 'GET',
           headers: {
             Authorization: `Bearer ${token}`,
@@ -172,8 +229,11 @@ describe('Pinning APIs endpoints', () => {
         'bafybeia45bscvzxngto555xsel4gwoclb5fxd7zpxige7rl3maoleznswu', // PinError
         'bafybeiaiipiibr7aletbbrzmpklw4l5go6sodl22xs6qtcqo3lqogfogy4' // Not exists
       ]
-
-      const url = new URL(`${baseUrl}?cid=${cids.join(',')}`).toString()
+      const opts = new URLSearchParams({
+        cid: cids.join(','),
+        status: 'failed,queued,pinning,pinned'
+      })
+      const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
         url, {
           method: 'GET',
@@ -191,7 +251,8 @@ describe('Pinning APIs endpoints', () => {
 
     it('filters case sensitive exact match on name', async () => {
       const opts = new URLSearchParams({
-        name: 'ReportDoc.pdf'
+        name: 'ReportDoc.pdf',
+        status: 'pinned'
       })
       const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
@@ -212,7 +273,8 @@ describe('Pinning APIs endpoints', () => {
     it('filters case insensitive partial match on name', async () => {
       const opts = new URLSearchParams({
         name: 'image',
-        match: 'ipartial'
+        match: 'ipartial',
+        status: 'failed,queued,pinning,pinned'
       })
       const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
@@ -252,7 +314,7 @@ describe('Pinning APIs endpoints', () => {
       assert.strictEqual(data.results[0].pin.name, 'FailedPinning.doc')
     })
 
-    it('filters pins by multiple status', async () => {
+    it('filters pins by multiple statuses', async () => {
       const opts = new URLSearchParams({
         status: 'queued,pinning'
       })
@@ -274,7 +336,8 @@ describe('Pinning APIs endpoints', () => {
 
     it('filters pins created before a date', async () => {
       const opts = new URLSearchParams({
-        before: '2021-07-01T00:00:00.000000Z'
+        before: '2021-07-01T00:00:00.000000Z',
+        status: 'failed,queued,pinning,pinned'
       })
       const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
@@ -289,12 +352,14 @@ describe('Pinning APIs endpoints', () => {
       assert(res, 'Server responded')
       assert(res.ok, 'Server response is ok')
       const data = await res.json()
+      assert.strictEqual(data.results.length, 1)
       assert.strictEqual(data.count, 1)
     })
 
     it('filters pins created after a date', async () => {
       const opts = new URLSearchParams({
-        after: '2021-07-15T00:00:00.000000Z'
+        after: '2021-07-15T00:00:00.000000Z',
+        status: 'failed,queued,pinning,pinned'
       })
       const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
@@ -309,12 +374,14 @@ describe('Pinning APIs endpoints', () => {
       assert(res, 'Server responded')
       assert(res.ok, 'Server response is ok')
       const data = await res.json()
+      assert.strictEqual(data.results.length, 1)
       assert.strictEqual(data.count, 1)
     })
 
     it('limits the number of pins returned for this user and includes the total', async () => {
       const opts = new URLSearchParams({
-        limit: '3'
+        limit: '3',
+        status: 'failed,queued,pinning,pinned'
       })
       const url = new URL(`${baseUrl}?${opts}`).toString()
       const res = await fetch(
@@ -384,10 +451,9 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, REQUIRED_CID)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorRequiredData.CODE)
+      assert.strictEqual(error.details, 'Instance does not have required property "cid".')
     })
 
     it('throws error if cid is invalid', async () => {
@@ -403,9 +469,8 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
       assert.strictEqual(error.details, INVALID_CID)
     })
 
@@ -431,9 +496,9 @@ describe('Pinning APIs endpoints', () => {
         })
       })
 
+      const data = await res.json()
       assert(res, 'Server responded')
       assert(res.ok, 'Server response ok')
-      const data = await res.json()
       assert.strictEqual(data.pin.cid, cid)
       assert.deepStrictEqual(data.pin.name, name)
       assert.deepStrictEqual(data.pin.origins, origins)
@@ -455,10 +520,9 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, INVALID_NAME)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/name: Instance type "number" is invalid. Expected "string".')
     })
 
     it('validates origins', async () => {
@@ -476,10 +540,9 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, INVALID_ORIGINS)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/origins: Instance type "number" is invalid. Expected "array".')
     })
 
     it('validates meta', async () => {
@@ -497,33 +560,9 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, INVALID_META)
-    })
-
-    it('validates meta values', async () => {
-      const res = await fetch(new URL('pins', endpoint).toString(), {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-          cid: 'bafybeibqmrg5e5bwhx2ny4kfcjx2mm3ohh2cd4i54wlygquwx7zbgwqs4e',
-          meta: {
-            app_id: 1
-          }
-        })
-      })
-
-      assert(res, 'Server responded')
-      assert.strictEqual(res.status, ERROR_CODE)
-      const data = await res.json()
-      const error = data.error
-      assert.strictEqual(error.reason, ERROR_STATUS)
-      assert.strictEqual(error.details, INVALID_META)
+      const error = await res.json()
+      assert.strictEqual(error.reason, PSAErrorInvalidData.CODE)
+      assert.strictEqual(error.details, '#/meta: Instance type "number" is invalid. Expected "object".')
     })
 
     it('error if user not authorized to pin', async () => {
@@ -545,7 +584,7 @@ describe('Pinning APIs endpoints', () => {
     })
 
     it('returns the pin request', async () => {
-      const sourceCid = 'QmVGv1UK8EvhD9KMHdKBB1LWadiYai6sY5GHL6h7MHRWzf'
+      const sourceCid = 'bafybeidhbtemubjbsbuhyai5oaebqf2fdrvhnshbkncyqpnoy2bl2mpt4q'
       const res = await fetch(new URL('pins', endpoint).toString(), {
         method: 'POST',
         headers: {
@@ -786,7 +825,7 @@ describe('Pinning APIs endpoints', () => {
       assert.deepEqual(res.status, 401)
     })
 
-    it('returns not found if the request does not exists', async () => {
+    it('returns not found if the request does not exist', async () => {
       const pinThatDoesNotExists = 'idThatDoesNotExists'
       const res = await fetch(new URL(`pins/${pinThatDoesNotExists}`, endpoint).toString(), {
         method: 'DELETE',
@@ -797,6 +836,9 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.deepEqual(res.status, 404)
+      const error = await res.json()
+      assert.deepEqual(error.reason, PSAErrorResourceNotFound.CODE)
+      assert.deepEqual(error.details, DATA_NOT_FOUND)
     })
 
     it('deletes the pin request', async () => {
@@ -878,40 +920,56 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert.equal(res.status, 404)
-      const { message } = await res.json()
-      assert.equal(message, 'Not Found')
+      const error = await res.json()
+      assert.equal(error.reason, PSAErrorResourceNotFound.CODE)
+      assert.equal(error.details, DATA_NOT_FOUND)
     })
 
     it('should delete the pin request and replace it', async () => {
+      // Create a pin request.
       const cid = 'bafybeid3ka3b3f443kv2je3mfm4byk6qps3wipr7wzu5uli6tdo57crcke'
-      const newCid = 'bafybeid4f2r3zpnkjqrglkng265ttqg6zbdr75dpbiwellvlpcxq7pggjy'
-
-      // Creates pin Requests
       const pinRequest = await createPinRequest(cid, token)
 
-      // It replaces it
-      const resR = await fetch(new URL(`pins/${pinRequest.requestId}`, endpoint).toString(), {
+      // Replace this pin request, and include extra data.
+      const newCid = 'bafybeid4f2r3zpnkjqrglkng265ttqg6zbdr75dpbiwellvlpcxq7pggjy'
+      const name = 'Replaced.pdf'
+      const origins = [
+        '/ip4/77.100.8.43/tcp/28253/p2p/12D3KooWGYUY2TCpPZsiaJfqs7V74mbSTgx4xNtBkRzkSGQjdaLp',
+        '/ip4/77.100.8.43/udp/28253/quic/p2p/12D3KooWGYUY2TCpPZsiaJfqs7V74mbSTgx4xNtBkRzkSGQjdaLp'
+      ]
+      const meta = { app_id: '99986338-1113-4706-8302-4420da6158aa' }
+
+      const replaceResponse = await fetch(new URL(`pins/${pinRequest.requestId}`, endpoint).toString(), {
         method: 'POST',
         headers: {
           Authorization: `Bearer ${token}`
         },
         body: JSON.stringify({
-          cid: newCid
+          cid: newCid,
+          name,
+          origins,
+          meta
         })
       })
 
-      assert(resR, 'Replace request did not respond')
-      assert(resR.ok, 'Replace request was not successful')
+      assert(replaceResponse, 'Replace request did not respond')
+      assert(replaceResponse.ok, 'Replace request was not successful')
+      assert.strictEqual(replaceResponse.status, 202)
+      const data = await replaceResponse.json()
+      assert.strictEqual(data.pin.cid, newCid)
+      assert.deepStrictEqual(data.pin.name, name)
+      assert.deepStrictEqual(data.pin.origins, origins)
+      assert.deepStrictEqual(data.pin.meta, meta)
 
-      const resG = await fetch(new URL(`pins/${pinRequest.requestId}`, endpoint).toString(), {
+      // Ensure the original pin request has been deleted.
+      const getResponse = await fetch(new URL(`pins/${pinRequest.requestId}`, endpoint).toString(), {
         method: 'GET',
         headers: {
           Authorization: `Bearer ${token}`
         }
       })
-
-      assert(resG, 'Get request did not respond')
-      assert.strictEqual(resG.status, 404, 'Pin request was not deleted')
+      assert(getResponse, 'Get request did not respond')
+      assert.strictEqual(getResponse.status, 404, 'Pin request was not deleted')
     })
 
     it('should not replace the same pin request', async () => {
@@ -929,8 +987,8 @@ describe('Pinning APIs endpoints', () => {
 
       assert(res, 'Server responded')
       assert(!res.ok)
-      assert.equal(res.status, 400)
-      const { error } = await res.json()
+      assert.equal(res.status, ERROR_CODE)
+      const error = await res.json()
       assert.equal(error.details, INVALID_REPLACE)
     })